packages/img-to-thumb/img-to-thumb/requirements.txt
boto3==1.26.60
requests==2.31.0
Pillow==9.2.0
